Content Aware Resizing
Content-aware resizing mainly uses an energy function to define the importance of pixels. The pixels with low energy are less important and can be eliminated when resizing the image. Figure 5 show the energy function of the Dolphin image, the pixels with large energy are white while those of small energy are black. Observe that the Dolphin pixels are the most important pixels with the largest energy, so it should not be removed by the Content-aware resizing.